coral_publisher/requirements.txt
# Coral Dev Board dependencies
# Install pycoral separately per: https://coral.ai/docs/setup/
# pip install --extra-index-url https://google-coral.github.io/py-packages/ pycoral~=2.0
opencv-python-headless>=4.8
pyzmq>=25.0
msgpack>=1.0
numpy>=1.24
